:art: Add config file
