+++ utils/config.go
@@ -1,0 +1,34 @@
+package utils
+
+import (
+	"os"
+	"time"
+
+	"github.com/charmbracelet/log"
+	"github.com/joho/godotenv"
+	"golang.org/x/oauth2"
+	"golang.org/x/oauth2/google"
+)
+
+var AppConfig oauth2.Config
+var StartTime time.Time
+
+func init() {
+	StartTime = time.Now()
+	err := godotenv.Load()
+	if err != nil {
+		log.Error(err)
+	}
+	log.Info("Loaded .env file")
+
+	AppConfig = oauth2.Config{
+		ClientID:     os.Getenv("GOOGLE_CLIENT_ID"),
+		ClientSecret: os.Getenv("GOOGLE_CLIENT_SECRET"),
+		Endpoint:     google.Endpoint,
+		RedirectURL:  os.Getenv("GOOGLE_REDIRECT_URI"),
+		Scopes: []string{
+			"https://www.googleapis.com/auth/userinfo.email",
+			"https://www.googleapis.com/auth/userinfo.profile",
+		},
+	}
+}
