+++ utils/utils.go
@@ -2,21 +2,16 @@ package utils
 
 import (
 	"time"
+
 	"github.com/gofiber/fiber/v2"
 )
-
-var startTime time.Time
-
-func init() {
-	startTime = time.Now()
-}
 
 func RootFunction(c *fiber.Ctx) error {
 	return c.JSON(fiber.Map{"message": "SRMKZILLA Recruitments-23 server running"})
 }
 
 func HealthCheck(c *fiber.Ctx) error {
-	return c.JSON(fiber.Map{"message": "OK", "uptime": time.Since(startTime).String(), "timestamp": time.Now()})
+	return c.JSON(fiber.Map{"message": "OK", "uptime": time.Since(StartTime).String(), "timestamp": time.Now()})
 }
 
 func HandleRoot(c *fiber.Ctx) error {
